requirements.txt
sqlalchemy==1.3.17
flask==1.1.2
flask_sqlalchemy==2.4.3
werkzeug==0.16.1
requests==2.23.0
bs4==0.0.1
eng-dictionary==0.0.3
PyYaml==5.3.1
mysql-connector-python-rf==2.2.2
gunicorn==20.0.4

